Fixed issue#19: Created a dummy query in FileConnectortype.validateConfig() and appended "additional where clause" in this query. Tested the syntax of the resultant query. If any error/exception occurs then an error message "Where Clause is syntactically incorrect" is logged and displayed on the configuration page.
This needed to change the resourcebundle file to add a new 
key i.e. additional_where_clause

Although the defect explation is only restricted to a single 
type of query i.e. related to folder, but query can be of any
type and can be syntactically wrong. So I think that here 
we need to validate "additional where clause" against
syntactical errors, instead of concentrating on only the 
folder kind of where clause.

Since FileNet SQL query processor allows the incorrect 
folder, so this cannot be fixed but, yes the syntax of the
"where clause" can be validated.

So this bug is fixed by the validation of "where clause".