FileNet 3.5 early feedback changes :
Issue 1.	There is no way to differentiate between the delete feeds of the documents matching the  additional delete clause and the documents which have already been deleted from the repository.
 Updated FileDocumentList.java,FileTraversalManager.java  files for the log messages accordingly.
 
Issue 2.	There is no validation if the user puts the same additional where clause and additional delete clause.
 Added validation for the same in FileConnectorType.java
Updated FileConnectorResources.properties for additional error messages.

Issue 3.	If the additional clause does not contain the validation “ContentSize IS NOT NULL”, it crawls all the documents and while searching the docs which have NULL content size throws exception.
 Added validation for the same in FileConnectorType.java