Unable to add a FileNet connector to the GSA (Case #02592094)
Fixes:
Customer was not able to persist the FileNet connector configuration.  This fix
is to make changes to the connection such that the Domain object will be fetched 
from the server before the connector fetches the object store.  In the prior 
version, the connector retrieves the Domain object locally which may not be 
available at the time.

https://codereview.appspot.com/6883045/