Split FileAuthorizationManager for testability and readability.
Leave the original class focused on the multi-threaded authorization,
and add a new FileAuthorizationHandler class using the FileNet API to
authorize a single document. (The original AuthorizationHandler class
for the worker Runnables is renamed AuthorizationTask.) Each class has
a clean focus, and we can easily test the threads without FileNet.

There are obvious cleanups to do (which three months later I no longer
remember), but I wanted to minimize the diffs in this change.

Code review: https://codereview.appspot.com/40980051/