Adds Recursive Security Folder Update
Recursively adds an AclDocument for each descendant Document of the
Folder to the list.

Changing the ACL on a folder changes the ACL on all its descendant
subfolders, yet the last modified time of those subfolders does not
get updated, so they don't show up in the query of modified folders.
Therefore the SecurityFolderTraverser must recursively update the
ACLs for all descendant subfolders. However, the connector does not
send folder ACLs to the GSA. It sends a copy of them with the docs
that inherit from those security folders. Therefore, we resend ACLs
for all documents who directly or indirectly inherit from the original
folder whose ACL changed.

Code Review: http://codereview.appspot.com/239890043