Fix FileConnectorTypeTest to use the correct properties.
is_public and authentication_type were removed long ago, and
check_marking and delete_additional_where_clause were added.

The TestConnection configuration for some properties was unnecessary
(since the needed values are not environment-specific):

    additional_where_clause
    wrong_additional_where_clause
    where_clause_error_message
    additional_delete_where_clause
    incorrectDisplayURL
    workplace_url__error_message
    repeatedSlashContentEngineURL

The additional_delete_where_clause property was unused.

The value of incorrectDisplayURL floating around for the internal
servers was actually the same value as displayUrl, and a valid
Workplace URL. The test passed because of a try statement that caught
the NullPointerException when validation succeeded, printed a bogus
"Incorrect URL" message to stdout, and exited.

testRepeatedSlashContentEngineURL did not use the
repeatedSlashContentEngineURL property at all, but reused the uri
property, so the test was simply a duplicate of
testValidateConfigWithBlankWhereClause.

Related changes:

* Extract the duplicate code into a setUp method, so that the test
  methods are much shorter (where these errors would have been more
  obvious). Also improved the error messages when assertions fails, to
  include the ConfigResponse message.

* Update google-enterprise-connector-file.properties.sample to remove
  the obsolete properties, and to add the missing domain, docId3, and
  docId4 properties.

Related changes in FileConnectorType:

* Log the actual URL validation exception. I did not remove the
  log-and-throw anti-pattern because I had enough with the first can
  of worms, so I wanted to preserve the downstream exceptions as-is.

* Do not catch Throwable (philosphical issue, not an actual problem).

Code review: https://codereview.appspot.com/233670043/