Remove IMarking and IActiveMarkingList in favor of FileNet interfaces.
The marking interfaces have no mocks (but could and should) and were
mostly straightforward wrappers of the FileNet interfaces. The
exception was that IActiveMarkingList was a list of IMarking rather
than IActiveMarking (which did not exist). The unwrapping of the
ActiveMarking to a Marking now happens in
MarkingPermissions.authorize.

Add TODOs to cleanup the use of markings in FileAuthorizationHandler.

This is part of merging fnapi-client and fnapi-core.

Code review: https://codereview.appspot.com/238510043/