Remove ISecurityTemplate and FnSecurityTemplate in favor of SecurityTemplate.
Replace Iterable<ISecurityTemplate> with the native SecurityTemplateList.

Catch EngineRuntimeException when handling security policies and
templates. FnSecurityTemplate mapped these to RepositoryException, but
FnSecurityPolicy did not, so these could previously have been thrown
from getDocumentList in the traverser.

Extract all logic from FnSecurityPolicy. The class is harder to remove
than it looks, since it's created by factories, but this is a step in
that direction.

Remove unused throws clauses from FnSecurityPolicy methods (they were
declared but not used).

Add a test that get_SecurityTemplates does not return null. It was
mishandled previously and would have thrown a NPE from the for loop in
SecurityPolicyTraverser. Since it can't happen, simplify the logic and
just handle empty lists. The test is in FnObjectFactoryTest. It does
involve the factories (stretching for justification...). More
importantly, there is no FnSecurityPolicyTest, and I was reluctant to
add one since FnSecurityPolicy has no logic and will be removed, and
SecurityPolicyTraverserTest has no access to the ObjectStore we need.

Code review: https://codereview.appspot.com/241740043/