Switch to FileNet API get_MemberOfGroups method in IUser.
This is part one of removing IUser.

Rename FileUserMock to UserMock for consistency. It will change more
in the next CL, so renaming it here so that Git follows the change.

Remove an optimization in Permissions that I never liked, which
depends on having the group names in a set. FnUser had a map from
group.get_Name() -> group, and we check various names of the group
(name, email, DN) against the grantee, but there was an optimization
to first check name against the map key set, before looping over the
groups to check the others.

Code review: https://codereview.appspot.com/244080043/