Remove IObjectSet and FnObjectList.
This is a large change but it's conceptually simple. Now that
MockBaseObject is gone and we have Documents in MockObjectStore, we
can postpone wrapping the search results in FnBaseObjects from
FnSearch.execute to FileDocumentList.addToList. While in flight, the
search results are now IndependentObjects in an IndependentObjectSet
instead of FnBaseObjects in an FnObjectList.

Both FnSearch methods now throw EngineRuntimeException.

Code review: https://codereview.appspot.com/245410043/