Smoke test SQL queries against H2.
The H2 tables are empty, and the mock results are returned as before.
The queries are simply executed against H2 to test for SQL syntax
errors.

For FileDocumentTraverser, the EasyMock mocks are carrying more
weight. Specifically, we expect different results against the same
table in different queries, and we are examining the queries
themselves. With non-empty tables we could do that, but with empty
tables, I just separately smoke test the queries, which were handily
captured by the mocks, directly from the test.

Copied JdbcFixture from the Documentum adaptor, which was refined from
the v3 versions.

Code review: http://codereview.appspot.com/240630043/