Remove FileDeletionObject.
This is a prelude to removing IBaseObject.

FileDeletionObject is a marker interface that simply delegates to the
underlying IBaseObject, but that will be more difficult when we
replace IBaseObject with IndependentObject, which has many more
methods. Instead, the objects in the three IObjectSets are wrapped in
a new non-delegating DocumentWrapper class that has an explicit type
for distinguishing added or updated documents from deleted documents
returned by the additional delete query. (DeletionEvents continue to
be identified using instanceof.)

The complicated changes are in the tests.

The tests started with the standard three sets of objects, but poured
them all into a map for the object store, illicitly wrapping the
custom deletions in FileDeletionObject (even before passing them to
FileDocumentList, which would wrap them again). Then the three sets
were extracted from the map by checking for instances of DeletionEvent
and FileDeletionObject.

This change flips this, so that we always start with three IObjectSets
containing unwrapped IBaseObjects, as in production. Then we put those
into the object store map, before passing the sets to FileDocumentList.
(The object store is used under the covers in FileDocument as before.)

Test notes:

* The sorting tests had only one non-empty object set, for documents.
  I have made this obvious, as a natural part of this change.

* newObjectStore takes a varargs array of IObjectSets and constructs
  the map internally.

* Merge the two testUnreleasedNextDocument methods.

* Remove the unused String[] overload of generateObjectMap and both
  overloads of generateCustomDeletion.

* getCustomDeletions (formerly getCustomDeletion) is an alias of
  getDocuments maintained for readability.

Other changes:

* Added two forgotten @Test annotations.

Code review: https://codereview.appspot.com/248880043/