Verify mocks in the test methods rather than an @After method.
Verifying the mocks in an @After method means that if a test throws an
exception before the test finishes, then there is likely to be an
unexercised mock. Verification in the @After method will fail,
dropping the root cause of the test failure irrevocably on the floor.

I kept the mechanism to save and verify all the mocks, but each test
is responsible for calling verifyAll(). A modified @After method makes
sure that you *did* verify the mocks in the tests when necessary,
which is a coding time failure, not one that will creep in later.

Code review: https://codereview.appspot.com/247040043/