Implement MockBaseObject on Document and DeletionEvent mocks.
This is a prelude to removing IBaseObject.

There are no changes to prod or test code, only to the implementation
of MockBaseObject itself.

Add comments for two existing behaviors that were uncovered by this
change. First, the sort in FileDocumentList must be stable. Second,
the FileDocumentList tests depend in a terrible way on the erroneous
assumption that the VersionSeries ID is the same as the ID of the
Document or DeletionEvent that refers to it.

Code review: https://codereview.appspot.com/249780043/