Add newly created MockBaseObjects to the MockObjectStore.
Ironically, this is a prelude to removing MockBaseObject. This change
sets the stage to mock DeletionEvent.getObjectStore in ObjectMocks,
which is required (along with some other things) to remove
MockBaseObject.isReleasedVersion.

Instead of adding MockBaseObjects in bulk from the IObjectSets later,
add them to the object store when they are created.

Replace a separate IBaseObject mock method in FileDocumentTraverserTest
and the PartialObjectStore class in in TraverserFactoryFixture with
ObjectMocks methods.

Code review: https://codereview.appspot.com/249050043/