Remove MockBaseObject.isReleasedVersion.
This is a shortcut to where I was trying to go. Instead of pushing all
the FileNet API object mocks into the object store, which requires
changing everything, I simply focused on implementing what was needed
for FnBaseObject.isReleasedVersion for DeletionEvents. Namely, the
mock DeletionEvent should return a mock ObjectStore [sic; N.B., not a
mock IObjectStore or MockObjectStore] that may or may not contain the
corresponding mock VersionSeries, depending on whether the
DeletionEvent is for the released version.

We still eventually want to have a mock ObjectStore with all the
object mocks, so this shortcut does not replace the other two changes
(putting objects into the object store early so we can implement
DeletionEvent.getObjectStore, and splitting the DeletionEvent ID and
VersionSeries ID, so both can be in the object store as needed).

Related changes:

* In FnBaseObject.isReleasedVersion, catch EngineRuntimeException
  instead of Exception. Since the code catches all Exceptions and
  returns false, debugging test failures was difficult.

Code review: http://codereview.appspot.com/247160043/