Remove more unused methods and type checks in the IBaseObject hierarchy.
Move methods called only on IBaseObject and not IDocument or
IVersionSeries to FnBaseObject, of which MockBaseObject is a subclass.
These are getModifyDate, getVersionSeriesId, and isReleasedVersion.

Corollary: All the objects in an IObjectSet and referenced in
           FileDocumentList are FnBaseObjects.

Replace isDeletionEvent, which was called once in FileDocumentList, by
extending ObjectType to distinguish between custom deleted Documents
and DeletionEvents.

All the objects in a MockObjectStore are MockBaseObjects.

Corollary: All constructor args to MockDocument and MockVersionSeries
           are MockBaseObjects.

In MockBaseObject, remove unused getProperty method.

Other changes:

* In MockDocument, use PropertyNames in place of ad hoc constants.

Code review: https://codereview.appspot.com/241570043/