Combine ObjectWrapper and FnBaseObject into SearchObject.
Move FnBaseObject out of the api package, do not implement IBaseObject,
and rename it to SearchObject.

Merge ObjectWrapper (including the ObjectType enum and compareTo method)
into the renamed SearchObject.

FnBaseObject was not really used outside of FileDocumentList or as an
IBaseObject. It appeared in FnObjectStore in else blocks that are never
executed, and in MockObjectStore in what essentially was an existence
test. I made the former errors and added a containsObject method to
MockObjectStore.

Code review: http://codereview.appspot.com/243630043/