Make ISearch.execute a thinner wrapper on SearchScope.fetchObjects.
SearchScope is a final class that uses internal classes to talk to the
FileNet repository. So we can't remove ISearch and FnSearch completely
and mock SearchScope. Instead, combine ISearch and FnSearch as a
mockable SearchWrapper class, and move closer to the actual
fetchObjects method signature.

We could expose SearchSQL, but since we always use full query strings,
it does not add value for the callers. So the fetchObjects wrapper
just takes a String instead of a SearchSQL, for convenience.

Helper fields in SearchWrapper provide more self-documentation for the
parameters, and the arguments have been tweaked. The smoke tests in
FileConnectorType don't need to fetch much data, we now use the
batchHint for the suggested page size instead of 100, and the
difference between dereferenceObjects and noFilter is hopefully more
clean than the former values of 1 and 0.

Other changes:

* Removed pointless doc comments and an unused throws clause in FileSession.

Code review: http://codereview.appspot.com/246350043/