Clean up authZ checkMarkings and hasMarkings.
Clean up the names and check the config before looking for marking sets
on the Document class.

FileAuthorizationManager and FileAuthorizationHandler had three
different variables related to marking sets, all called checkMarkings.
Now checkMarkings refers to the connector configuration and
authorizeMarkings is the return value from hasMarkings(), passed back
to FileAuthorizationHandler.authorizeDocid.

There was a cache-like feature in reassigning checkMarkings in
hasMarkings(), but it didn't work. Add a TODO to investigate the
correct caching requirements.

Do not check the Document class for a marking set if checkMarkings is
false, since we will not authorize against the marking set in that
case.

Code review: https://codereview.appspot.com/248600043/