Refactor FileNet DependentEngineList collections.
Modify IObjectStore.getPropertyDefinitions to return a
PropertyDefinitionList rather than an Iterator, and get the Iterator
in the caller. This is part of moving things out of the api package.

Replace the large AccessPermissionListMock with the simpler style used
by ActiveMarkingListMock and others. Call Collections.shuffle directly.

Rename EngineSetMocks to EngineCollectionMocks and move all the
DependentEngineList mocks there.

Code review: https://codereview.appspot.com/254930043/