Clean up obsolete properties and default values.
Add useIDForChangeDetection and is_public to connectorDefaults.xml.
Move them into the commented-out section in connectorInstance.xml.

Remove EncryptedPropertyPlaceholderConfigurer entries for properties
that have always been on the configuration form, or which are not
config form properties.

Remove object_factory as a config form property. The entry in
connectorDefaults.xml did not use a placeholder and its use is purely
internal.

Remove the default value of the deprecated db_timezone property.

Remove the generation of hidden inputs for obsolete and unknown keys
in the config map.

Add mock tests for validateConfig.

Other changes:

* Make password inputs 50 columns like text inputs.

* Inline appendEndRow, appendStartHiddenRow, and parts of
  appendStartRow, now renamed to appendLabel.

Code review: https://codereview.appspot.com/254020043