Do not feed ACL for document if it has an active marking
The connector does not currently implement ACLs for marking sets,
so if a document has an active marking, then the ACLs do not
accurately represent the security of the document.

As a stop-gap, if a document has an active marking,
we drop the ACL on the floor (as with pushAcls=false)
so that connector authZ will be used for that document.

This change borrows code from FileAuthorizationHandler
to determine if we should check for active markings.
Unlike FileAuthorizationHandler, this changes only checks
the FileNet Document Class for properties that may have
markings once, and caches the result.

If the additional WHERE clause starts with SELECT,
then the config may not be using the Document class.
In that case, we just check for active markings on
every document.

Code Review: http://codereview.appspot.com/254490043